requirements.txt
python-dotenv==1.0.1
pytest==8.2.2
numpy==1.26.4
torch==2.3.1
torchvision==0.18.1
opencv-python==4.10.0.84
pillow==10.3.0
mediapipe==0.10.14
scikit-learn==1.5.0